requirements.txt
flask
python-dotenv
PyPDF2
python-docx
browser-use
langchain-google-genai
playwright